bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
uvicorn==0.30.0
fastapi==0.111.0
apscheduler==3.11.0
pytest
pytest-asyncio
watchdog
pytest-mock
